Created a function that will populate the show variable that is sued by the DOM to display shows. populateShows takes in the form value from the show type. It uses the check boxes to generate how many films/ series should be added to the variable shows. If movie is ticked it'll get 10 movies, if only tv is ticked it'll get 10 shows. If both/ none are ticked it'll get a random amount of movies.
Next step is to create functionality for other filters. For the IMDB rating this will be done by calling the discover api with the slider values passed in, we will then get a random page number and pick a show from it. We will need an initial api call to find out how many pages there are for movie and tv. The entire finding latest will be removed and replaced with page number getting for discover api